use config in bird
