+++ bird/config.go
@@ -1,0 +1,16 @@
+package bird
+
+// Birdwatcher Birdc Configuration
+
+type StatusConfig struct {
+	ReconfigTimestampSource string `toml:"reconfig_timestamp_source"`
+	ReconfigTimestampMatch  string `toml:"reconfig_timestamp_match"`
+
+	FilteredFields []string `toml:"filtered_fields"`
+}
+
+type BirdConfig struct {
+	Listen         string
+	ConfigFilename string `toml:"config"`
+	BirdCmd        string `toml:"birdc"`
+}
